+++ xdr/asset_test.go
@@ -1,7 +1,9 @@
-package xdr
+package xdr_test
 
 import (
 	"testing"
+
+	. "github.com/stellar/go/xdr"
 
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
@@ -109,22 +111,6 @@ func TestStringCanonical(t *testing.T) {
 
 	asset = MustNewCreditAsset("USD", "GBRPYHIL2CI3FNQ4BXLFMNDLFJUNPU2HY3ZMFSHONUCEOASW7QC7OX2H")
 	require.Equal(t, "USD:GBRPYHIL2CI3FNQ4BXLFMNDLFJUNPU2HY3ZMFSHONUCEOASW7QC7OX2H", asset.StringCanonical())
-}
-
-func TestTrimRightZeros(t *testing.T) {
-	require.Equal(t, []byte(nil), trimRightZeros(nil))
-	require.Equal(t, []byte{}, trimRightZeros([]byte{}))
-	require.Equal(t, []byte{}, trimRightZeros([]byte{0x0}))
-	require.Equal(t, []byte{}, trimRightZeros([]byte{0x0, 0x0}))
-	require.Equal(t, []byte{0x1}, trimRightZeros([]byte{0x1}))
-	require.Equal(t, []byte{0x1}, trimRightZeros([]byte{0x1, 0x0}))
-	require.Equal(t, []byte{0x1}, trimRightZeros([]byte{0x1, 0x0, 0x0}))
-	require.Equal(t, []byte{0x1}, trimRightZeros([]byte{0x1, 0x0, 0x0, 0x0}))
-	require.Equal(t, []byte{0x1, 0x2}, trimRightZeros([]byte{0x1, 0x2}))
-	require.Equal(t, []byte{0x1, 0x2}, trimRightZeros([]byte{0x1, 0x2, 0x0}))
-	require.Equal(t, []byte{0x1, 0x2}, trimRightZeros([]byte{0x1, 0x2, 0x0, 0x0}))
-	require.Equal(t, []byte{0x0, 0x2}, trimRightZeros([]byte{0x0, 0x2, 0x0, 0x0}))
-	require.Equal(t, []byte{0x0, 0x2, 0x0, 0x1}, trimRightZeros([]byte{0x0, 0x2, 0x0, 0x1, 0x0}))
 }
 
 var _ = Describe("xdr.Asset#Equals()", func() {
